Store money as integer cents instead of float64
Represent every monetary amount as int64 cents end-to-end (domain, services,
DTOs) to eliminate floating-point rounding. Split calculations now use a
largest-remainder distribution so the parts always sum exactly to the total
(e.g. 1000¢/3 -> 334/333/333). The API contract changes: amounts are integer
cents in requests and responses. Column types drop numeric(10,2) for bigint.

NOTE: requires the money tables to be recreated (no in-place data migration);
deploy is coordinated with the frontend cents change.

+++ internal/domain/balance.go
@@ -2,11 +2,11 @@ package domain
 
 type UserBalance struct {
 	UserID uint
-	Amount float64 // Positive means they should receive money, Negative means they owe money
+	Amount int64 // cents; positive means they should receive money, negative means they owe
 }
 
 type Debt struct {
-	FromUserID uint    `json:"from_user_id"`
-	ToUserID   uint    `json:"to_user_id"`
-	Amount     float64 `json:"amount"`
+	FromUserID uint  `json:"from_user_id"`
+	ToUserID   uint  `json:"to_user_id"`
+	Amount     int64 `json:"amount"` // cents
 }
